fix: qdrant search API, youtube-transcript-api v1.x, gemini rate limits
- VideoData.duration: int -> float (yt-dlp returns 60.465, not 60)
- Video ORM: duration Integer -> Float column
- StructureSegment.segment: Literal -> str (LLM may return lowercase)
- TimelineView: normalise segment names to lowercase before lookup,
  replace invisible bg-white/N colors with purple/blue/emerald/orange
- ChatPanel: fix citation tags wrapping, add Sources label, improve layout
- formatDuration: use Math.floor() to handle fractional seconds
- types/index.ts: StructureSegment.segment: string (not strict Literal)

+++ backend/app/models/db.py
@@ -25,7 +25,7 @@ class Video(Base):
     likes: Mapped[int | None] = mapped_column(Integer)
     comments: Mapped[int | None] = mapped_column(Integer)
     engagement_rate: Mapped[float | None] = mapped_column(Float)
-    duration: Mapped[int | None] = mapped_column(Integer)  # seconds
+    duration: Mapped[float | None] = mapped_column(Float)  # seconds (float from yt-dlp)
     upload_date: Mapped[str | None] = mapped_column(String(20))
     hashtags: Mapped[list | None] = mapped_column(JSON)
     transcript: Mapped[str | None] = mapped_column(Text)
